Change associations Strategy for hashtags and Categories, fixed bug for wrong email login (error message), merge Category crud and hashtag crud to Recipe Crud,

+++ src/app/services/recipeLike.service.js
@@ -1,8 +1,17 @@
 import { RecipeLikes } from '../models/associations';
+import { Conflict } from "../errors";
 
 class RecipeLikeService {
 
     async like({ user_id, recipe_id }) {
+        const like = await RecipeLikes.findOne({
+            where: {
+                user_id,
+                recipe_id
+            }
+        });
+        if (like) { throw new Conflict('already liked') }
+
         return await RecipeLikes.create({
             user_id,
             recipe_id
@@ -10,6 +19,14 @@ class RecipeLikeService {
     }
 
     async disLike({ user_id, recipe_id }) {
+        const like = await RecipeLikes.findOne({
+            where: {
+                user_id,
+                recipe_id
+            }
+        });
+        if (!like) { throw new Conflict('already disLiked') }
+
         return await RecipeLikes.destroy({
             where: {
                 user_id,
